% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fn_mode}
\alias{fn_mode}
\title{Calculating mode value of a vector}
\usage{
fn_mode(x)
}
\arguments{
\item{x}{a vector of string or number}
}
\value{
The function returns the mode value of the input vector.
}
\description{
The function returns the mode of a vector. The vector can be of any datatype ie. numerical or categorical.
}
\examples{
fn_mode(c(1,2,3,1,4,1,7))
}
\author{
Arya Poddar <aryapoddar290990@gmail.com>
}
